+++ source/2_platform/rex_directx/include/rex_directx/rendering/dx_imgui_child_viewport.h
@@ -1,0 +1,17 @@
+#pragma once
+
+#include "rex_directx/rendering/dx_imgui_viewport.h"
+
+namespace rex
+{
+  namespace renderer
+  {
+    class ImGuiChildViewport : ImGuiViewport
+    {
+    public:
+
+    private:
+
+    };
+  }
+}
